+++ src/Ext/WeaponType/Body.h
@@ -42,8 +42,8 @@ public:
 		Valueable<AffectedHouse> CanTargetHouses;
 		Valueable<double> CanTarget_MaxHealth;
 		Valueable<double> CanTarget_MinHealth;
-		Nullable<float> CanTarget_MaxVeterancy;
-		Valueable<float> CanTarget_MinVeterancy;
+		Nullable<double> CanTarget_MaxVeterancy;
+		Valueable<double> CanTarget_MinVeterancy;
 		ValueableVector<int> Burst_Delays;
 		Valueable<bool> Burst_FireWithinSequence;
 		Valueable<bool> Burst_NoDelay;
@@ -117,7 +117,7 @@ public:
 			, CanTarget_MaxHealth { 1.0 }
 			, CanTarget_MinHealth { 0.0 }
 			, CanTarget_MaxVeterancy {}
-			, CanTarget_MinVeterancy { 0.0f }
+			, CanTarget_MinVeterancy { 0.0 }
 			, Burst_Delays {}
 			, Burst_FireWithinSequence { false }
 			, Burst_NoDelay { false }
